+++ projects/angular2-smart-table/src/lib/components/filter/filter-types/mselect-filter.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { DefaultFilter } from './default-filter';
 import { HttpClient } from "@angular/common/http";
 import { DropdownSettings } from 'angular2-multiselect-dropdown/lib/multiselect.interface';
+import {defaultStringInclusionFilter} from '../../../lib/data-source/local/local.filter';
 
 export interface Config {
     dropdownList: Array<any>,
@@ -11,12 +12,12 @@ export interface Config {
 
 @Component({
     selector: 'mselect-filter',
-    template: `<angular2-multiselect [data]="dropdownList" 
-    [(ngModel)]="selectedItems" 
-    [settings]="dropdownSettings" 
+    template: `<angular2-multiselect [data]="dropdownList"
+    [(ngModel)]="selectedItems"
+    [settings]="dropdownSettings"
     (onSelect)="onItemSelect($event)"
-    (onDeSelect)="OnItemDeSelect($event)" 
-    (onSelectAll)="onSelectAll($event)" 
+    (onDeSelect)="OnItemDeSelect($event)"
+    (onSelectAll)="onSelectAll($event)"
     (onDeSelectAll)="onDeSelectAll($event)">
     </angular2-multiselect>`
 })
@@ -47,7 +48,6 @@ export class MselectFilterComponent extends DefaultFilter implements OnInit {
             maxHeight: 200,
             position: 'top',
             autoPosition: true,
-
         };
         this.dropdownSettings = Object.assign(setting, config.dropdownSettings);
     }
@@ -55,12 +55,9 @@ export class MselectFilterComponent extends DefaultFilter implements OnInit {
     onfilterValues(cellValue: string, search: string, data: any, cellName: string) {
         if (search.indexOf(this.selector) != -1) {
             let searchArray = search.split(this.selector);
-            if (searchArray.indexOf(cellValue) != -1) {
-                return true;
-            }
-            return false;
+            return searchArray.indexOf(cellValue) != -1;
         }
-        return cellValue.toString().toLowerCase().includes(search.toString().toLowerCase());
+        return defaultStringInclusionFilter(cellValue, search, data, cellName);
     }
 
     onItemSelect(item: any) {